+++ src/FileEmbedLzma.hs
@@ -32,6 +32,7 @@ import Control.Monad.Trans.State.Strict (runState, state)
 import Data.Foldable                    (for_)
 import Data.Functor.Compose             (Compose (..))
 import Data.Int                         (Int64)
+import Data.List                        (sort)
 import Language.Haskell.TH
 import Language.Haskell.TH.Syntax       (qAddDependentFile)
 import System.Directory
@@ -52,6 +53,13 @@ import qualified Data.ByteString.Internal as BS.Internal
 
 import Language.Haskell.TH.Syntax (Bytes (..))
 #endif
+
+-- $setup
+-- >>> :set -XTemplateHaskell -dppr-cols=9999
+-- >>> import qualified Data.ByteString.Lazy as LBS
+-- >>> import qualified Data.ByteString as BS
+-- >>> import qualified Data.Text.Lazy as LT
+-- >>> import qualified Data.Text as T
 
 listRecursiveDirectoryFiles :: FilePath -> IO [(FilePath, LBS.ByteString)]
 listRecursiveDirectoryFiles = listDirectoryFilesF listRecursiveDirectoryFiles
@@ -155,11 +163,19 @@ embedPairs pairs = do
 -- -- is an embedded (no data-files!) equivalent of
 -- staticApp $ defaultFileServerSettings "static"
 -- @
+--
+--
+-- >>> $(embedRecursiveDir "example")
+-- [("/Example.hs","..."),("/example.txt","Hello from the inside.\n")]
+--
+-- >>> :t $(embedRecursiveDir "example")
+-- $(embedRecursiveDir "example") :: [(FilePath, BS.ByteString)]
+--
 embedRecursiveDir :: FilePath -> Q Exp
 embedRecursiveDir topdir = do
     pairs' <- runIO $ listRecursiveDirectoryFiles topdir
     for_ pairs' $ qAddDependentFile . fst
-    let pairs = makeAllRelative topdir pairs'
+    let pairs = sort (makeAllRelative topdir pairs')
     embedPairs pairs
 
 -------------------------------------------------------------------------------
@@ -167,6 +183,10 @@ embedRecursiveDir topdir = do
 -------------------------------------------------------------------------------
 
 -- | Embed a lazy 'Data.ByteString.Lazy.ByteString' from a file.
+--
+-- >>> :t $(embedLazyByteString "file-embed-lzma.cabal")
+-- $(embedLazyByteString "file-embed-lzma.cabal") :: LBS.ByteString
+--
 embedLazyByteString :: FilePath -> Q Exp
 embedLazyByteString fp = do
     qAddDependentFile fp
@@ -174,10 +194,18 @@ embedLazyByteString fp = do
     lazyBytestringE bsl
 
 -- | Embed a strict 'Data.ByteString.ByteString' from a file.
+--
+-- >>> :t $(embedByteString "file-embed-lzma.cabal")
+-- $(embedByteString "file-embed-lzma.cabal") :: BS.ByteString
+--
 embedByteString :: FilePath -> Q Exp
 embedByteString fp = [| LBS.toStrict |] `appE` embedLazyByteString fp
 
 -- | Embed a lazy 'Data.Text.Lazy.Text' from a UTF8-encoded file.
+--
+-- >>> :t $(embedLazyText "file-embed-lzma.cabal")
+-- $(embedLazyText "file-embed-lzma.cabal") :: LT.Text
+--
 embedLazyText :: FilePath -> Q Exp
 embedLazyText fp = do
     qAddDependentFile fp
@@ -188,5 +216,9 @@ embedLazyText fp = do
     [| LTE.decodeUtf8 |] `appE` lazyBytestringE bsl
 
 -- | Embed a strict 'Data.Text.Text' from a UTF8-encoded file.
+--
+-- >>> :t $(embedText "file-embed-lzma.cabal")
+-- $(embedText "file-embed-lzma.cabal") :: T.Text
+--
 embedText :: FilePath -> Q Exp
 embedText fp = [| LT.toStrict |] `appE` embedLazyText fp